Removed string field from `Lines`-iterator
We don't need it for the iterator itself, neither is there a way to get
the original string back from iterator, so why save it there? Also it is
already saved in `Split`, where it is actually needed by the algorithm. So
keeping the string in `Lines` would just double the memory footprint for
nothing.