Some fixes to ATA DMA driver
This is an attempt to get Redox to boot in VirtualBox using the ATA
driver in DMA mode. Some reads that were previously failing are now
working, but other ones that previously worked are now failing.

The main change here is that we send an ATA command for each PRD in the
PRDT, instead of a single command for the entire section to be read.

There is also a change to separate out the EOT u16 part in the Prd from
the size (the size can only be u16 anyways), and using the right bytes
for HDDEVSEL (the old bytes were for the 28-bit LBAs instead of the
48-bit LBAs that we use.