kernel: improve syscall argument handling
Currently, pointers passed as syscall arguments are passed unchecked to
the applicable handler. Most correctly verify that they are valid
userspace pointers, but some (pipe2, exec) do not, causing potential
security issues. This change moves (almost) all of the pointer checking
to the syscall dispatcher so that it's easy to verify that things are
being checked correctly.

The exceptions are the last argument to futex, which is only used for
some operations and so is checked by the handler only if it is used, and
exec, which has not been converted to the new ABI and would be a bit
more complicated to check. Until this is fixed, exec is vulnerable to a
userspace process being able to leak kernel data.