Implement capability mangament.
At this moment it is relatively useless due to the lack of control
syscalls, but nonetheless the data structures themself are implemented.

Every context has a map of schemes and associated capability sets. A
capability is a member of such set (or a supercapability of a member),
containing two fields, a byte sequence (data) and a type label (kind).

Putting all this together allows for checking for downgrading
capabilities.