prefix: Fix build breakage due to invalid prefix argument
Building a rust compiler for x86_64-unknown-redox using:

$ PREFIX_RUSTC=1 PREFIX_BINARY=0 make prefix

fails with a message citing 'could not canonicalize prefix path'. As
seen in rust/bootstrap/install.rs:

https://gitlab.redox-os.org/redox-os/rust/blob/redox/src/bootstrap/install.rs#L77

.. the expectation is that either the '--prefix' argument passed to
rust's configure script is a valid path or not specified at all (in
which case a correct default is used).

This patch uses the former approach resulting in a correct build.