mk/kernel.mk: aarch64: Use full path to target JSON spec
In that JSON spec file, we now use target-features to specify the use of
the tpidr_el1 instruction for kernel side TLS accesses. This means we
can step aside from the unfortunate LLVM requirement to use the kernel
code model in order for tpidr_el1 instructions to be emitted.

In addition, we suppress generation of NEON instructions for the kernel.