Allow local sources
This makes the source section of recipe.toml optional, so that one can
simply create a recipe with a build recipe and a source folder and not
have to worry about git at all just to get a program running in redox.

I also ran into a situation where a source directory without a git
repository in it caused cookbook to reset it's own origin URL to the URL
specified in the recipe for that source directory. I added a check to
prevent that happening to anybody else.