Update Boost recipe
Boost depends on some missing `wchar.h` functions which are wrapped and
exposed by C++. Some of the libraries seem to either depend on `openat`
or depend on external libraries that depend on `openat`.

The main purpose of this patch is to update the recipe to an almost
working state until those issues are resolved.